Home

IntervalMatrices.jl

IntervalMatrices is a Julia package to work with matrices that have uncertain parameters.

Features

An IntervalMatrix type.

Arithmetics between scalars, closed intervals and interval matrices.

Overapproximation and underapproximation routines for the exponential of an interval matrix at different orders.

Library Outline

IntervalMatrices — Module.

Main module for IntervalMatrices.jl – a Julia package to work with matrices that have uncertain parameters.

source

Types

Abstract interval operators

Interval matrix

Methods

Hulls