Types

Types

This section describes systems types implemented in IntervalMatrices.jl.

Types

Abstract interval operators

Interval matrix

Interval-matrix-power wrapper

Abstract interval operators

IntervalMatrices.AbstractIntervalMatrix — Type.

AbstractIntervalMatrix{IT} <: AbstractMatrix{IT}

Abstract supertype for interval matrix types.

source

Interval matrix

IntervalMatrices.IntervalMatrix — Type.

IntervalMatrix{T, IT<:Interval{T}, MT<:AbstractMatrix{IT}} <: AbstractIntervalMatrix{IT}

An interval matrix i.e. a matrix whose coefficients are intervals. This type is parametrized in the number field, the interval type, and the matrix type.

Fields

mat – matrix whose entries are intervals

Examples

julia> A = IntervalMatrix([-0.9±0.1 0±0; 0±0 -0.9±0.1])
2×2 IntervalMatrix{Float64,Interval{Float64},Array{Interval{Float64},2}}:
 [-1, -0.799999]   [0, 0]
  [0, 0]          [-1, -0.799999]

An interval matrix proportional to the identity matrix can be built using the UniformScaling operator from the standard library LinearAlgebra. For example,

julia> using LinearAlgebra

julia> IntervalMatrix(Interval(1)*I, 2)
2×2 IntervalMatrix{Float64,Interval{Float64},Array{Interval{Float64},2}}:
 [1, 1]  [0, 0]
 [0, 0]  [1, 1]

The number of columns can be specified as a third argument, creating a rectangular $m × n$ matrix such that only the entries in the main diagonal, $(1, 1), (2, 2), …, (k, k)$ are specified, where $k = \min(m, n)$:

julia> IntervalMatrix(Interval(-1, 1)*I, 2, 3)
2×3 IntervalMatrix{Float64,Interval{Float64},Array{Interval{Float64},2}}:
 [-1, 1]   [0, 0]  [0, 0]
  [0, 0]  [-1, 1]  [0, 0]

julia> IntervalMatrix(Interval(-1, 1)*I, 3, 2)
3×2 IntervalMatrix{Float64,Interval{Float64},Array{Interval{Float64},2}}:
 [-1, 1]   [0, 0]
  [0, 0]  [-1, 1]
  [0, 0]   [0, 0]

An uninitialized interval matrix can be constructed using undef:

julia> m = IntervalMatrix{Float64}(undef, 2, 2);

julia> typeof(m)
IntervalMatrix{Float64,Interval{Float64},Array{Interval{Float64},2}}

Note that this constructor implicitly uses a dense matrix, Matrix{Float64}, as the matrix (mat) field in the new interval matrix.

source

Interval-matrix-power wrapper

IntervalMatrices.IntervalMatrixPower — Type.

IntervalMatrixPower{T}

A wrapper for the matrix power that can be incremented.

Fields

M – the original matrix

Mᵏ – the current matrix power, i.e., $M^k$

k – the current power index

Notes

The wrapper should only be accessed using the interface functions. The internal representation (such as the fields) are subject to future changes.

Examples

julia> A = IntervalMatrix([2.0..2.0 2.0..3.0; 0.0..0.0 -1.0..1.0])
2×2 IntervalMatrix{Float64,Interval{Float64},Array{Interval{Float64},2}}:
 [2, 2]   [2, 3]
 [0, 0]  [-1, 1]

julia> pow = IntervalMatrixPower(A);

julia> increment!(pow)
2×2 IntervalMatrix{Float64,Interval{Float64},Array{Interval{Float64},2}}:
 [4, 4]  [2, 9]
 [0, 0]  [0, 1]

julia> increment(pow)
2×2 IntervalMatrix{Float64,Interval{Float64},Array{Interval{Float64},2}}:
 [8, 8]  [-1, 21]
 [0, 0]   [-1, 1]

julia> get(pow)
2×2 IntervalMatrix{Float64,Interval{Float64},Array{Interval{Float64},2}}:
 [4, 4]  [2, 9]
 [0, 0]  [0, 1]

julia> index(pow)
2

julia> base(pow)
2×2 IntervalMatrix{Float64,Interval{Float64},Array{Interval{Float64},2}}:
 [2, 2]   [2, 3]
 [0, 0]  [-1, 1]

source